+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [v4.8-2021a] not yet released
 ### Added
+- Specific leap month queries in East Asian calendars [#913]
 - New methods for solar terms in East Asian calendars creating lists and operators [#928]
 
 ### Changed